color problems
